+++ src/main/java/container/application/ContainerAttribute.java
@@ -9,5 +9,6 @@ public enum ContainerAttribute {
     CODE,
     NAME,
     WEIGHT,
+    ISBOX,
     ISDELETED;
 }