+++ src/main/java/deliverynote/application/DeliveryNoteItemAttribute.java
@@ -1,9 +1,0 @@
-package deliverynote.application;
-
-/**
- * Represents the attributes for a delivery note item.
- */
-public enum DeliveryNoteItemAttribute {
-    CONTAINER,
-    QTY
-}