+++ CHANGELOG.md
@@ -7,11 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.46.0] - 2023-05-13
+## [0.49.0] - 2023-05-22
 
 ### Added
 
 - Added feature to generate a delivery note.
+
+## [0.48.0] - 2023-05-21
+
+### Added
+
+- Add a weighings table.
+
+## [0.47.0] - 2023-05-20
+
+### Added
+
+- Add distinction between pallet and box on the container entity.
+
+## [0.46.0] - 2023-05-19
+
+### Removed
+
+- Removed the price from the product.
+>>>>>>> origin/main
 
 ## [0.45.0] - 2023-05-07